+++ NEWS.md
@@ -1,4 +1,4 @@
-## 2.2.0 2021-09-xx 
+## 2.2.0 - Released 
 
 ### Changes
 * [MODQM-134](https://issues.folio.org/browse/MODQM-134) - Change dataType to have common type for MARC related subtypes
@@ -9,6 +9,7 @@
 * [MODQM-142](https://issues.folio.org/browse/MODQM-142) - Backend: Edit MARC Holdings via quickMARC
 * [MODQM-150](https://issues.folio.org/browse/MODQM-150) - MARC holdings - 008 rules
 * [MODQM-151](https://issues.folio.org/browse/MODQM-151) - MARC holdings - Leader rules
+* [MODQM-154](https://issues.folio.org/browse/MODQM-154) - MARC authority - Leader rules
 
 ## 2.1.0 - Released